moved most of payment configuration from PaymentComponent to PaymentService
