+++ CKODemoShop/ClientApp/src/app/components/payment-method-form/payment-method-form.component.html
@@ -2,10 +2,10 @@
   <div class="payment-setup" formGroupName="source">
     <mat-radio-group class="example-radio-group" formControlName="type" required>
       <mat-radio-button class="example-radio-button" *ngFor="let paymentMethod of paymentMethods" [value]="paymentMethod.type" [disabled]="!sourceTypeSupportsCurrencyCountryPairing(paymentMethod.type)">
-        <span matTooltip="{{sourceTypeSupportsCurrencyCountryPairing(paymentMethod.type) ? '' : 'Billing Country and/or Currency mismatch'}}">{{paymentMethod.name}}</span>
+        <span matTooltip="{{sourceTypeSupportsCurrencyCountryPairing(paymentMethod.type) ? '' : 'Unsupported payment country and/or payment currency'}}">{{paymentMethod.name}}</span>
       </mat-radio-button>
     </mat-radio-group>
-    <mat-card *ngIf="paymentMethodRequiresAdditionalInformation">
+    <mat-card *ngIf="sourceFieldsCount > 1">
       <mat-card-header>
         <mat-card-title>{{paymentMethodName}}</mat-card-title>
         <mat-card-subtitle>Payment Information</mat-card-subtitle>
